@@ -99,7 +99,8 @@ def fixture_java_exe():
             mtap_jar = next(java_libs.glob('*-all.jar'))
         except (FileNotFoundError, StopIteration):
             raise ValueError('MTAP_JAR not found, is required for tests')
+    cp = ['-cp', str(mtap_jar) + ':' + str(pathlib.Path(__file__).parent / 'slf4j-simple-2.0.3.jar')]
     try:
-        return [str(pathlib.Path(os.environ['JAVA_HOME']) / 'bin' / 'java'), '-cp', str(mtap_jar)]
+        return [str(pathlib.Path(os.environ['JAVA_HOME']) / 'bin' / 'java')] + cp
     except KeyError:
-        return ['java', '-cp', str(mtap_jar)]
+        return ['java'] + cp
